% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConvergenceSpeed.R
\name{calcTrigonometricAreas}
\alias{calcTrigonometricAreas}
\title{Title}
\usage{
calcTrigonometricAreas(solver_traj, triangle, trapezoid)
}
\arguments{
\item{solver_traj}{}

\item{triangle}{}

\item{trapezoid}{}
}
\value{

}
\description{
Title
}
