% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_monitor_feature.R
\name{checkPlat}
\alias{checkPlat}
\title{Plateau Checker (only needed for dataloader)}
\usage{
checkPlat(solvertraj)
}
\arguments{
\item{solvertraj}{}
}
\value{

}
\description{
Plateau Checker (only needed for dataloader)
}
