% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlateauStats.R
\name{generateIntervalplot}
\alias{generateIntervalplot}
\title{Title}
\usage{
generateIntervalplot(data, incumbant = "incumbant", V1 = "V1")
}
\arguments{
\item{data}{}

\item{incumbant}{}

\item{V1}{}
}
\value{

}
\description{
Title
}
