% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConvergenceSpeed.R
\name{getConvSpeed_2}
\alias{getConvSpeed_2}
\title{Title}
\usage{
getConvSpeed_2(solver_traj, tri_area, trap_area)
}
\arguments{
\item{solver_traj}{}

\item{tri_area}{}

\item{trap_area}{}
}
\value{

}
\description{
Title
}
