% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultTrajStats.R
\name{getDefaultStats}
\alias{getDefaultStats}
\title{Title}
\usage{
getDefaultStats(solvertraj, solvertraj_copy = "", eff_real_stat = FALSE)
}
\arguments{
\item{solvertraj}{}

\item{solvertraj_copy}{}

\item{eff_real_stat}{}
}
\value{

}
\description{
Title
}
