% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKneeRatio.R
\name{getKneeRatio}
\alias{getKneeRatio}
\title{Title}
\usage{
getKneeRatio(solver_traj, VG_stats, plat_start_stats)
}
\arguments{
\item{solver_traj}{}

\item{VG_stats}{}

\item{plat_start_stats}{}
}
\value{

}
\description{
Title
}
