% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastXimprovementRatio.R
\name{getLastXimprovement}
\alias{getLastXimprovement}
\title{Title}
\usage{
getLastXimprovement(
  solver_traj,
  p_percent_observation = NULL,
  backwards = TRUE,
  from_to = FALSE,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{solver_traj}{}

\item{p_percent_observation}{}

\item{backwards}{}

\item{from_to}{}

\item{from}{}

\item{to}{}
}
\value{

}
\description{
Title
}
