% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitoringFeatures.R
\name{getMonitoringFeatures}
\alias{getMonitoringFeatures}
\title{Monitoring Featre main getter}
\usage{
getMonitoringFeatures(instance_ls, cutoff.time, scaling)
}
\arguments{
\item{instance_ls}{}

\item{cutoff.time}{}

\item{scaling}{}
}
\value{

}
\description{
Monitoring Featre main getter
}
