% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlateauStats.R
\name{getPlateauStats}
\alias{getPlateauStats}
\title{Title}
\usage{
getPlateauStats(solver_traj, successRatio)
}
\arguments{
\item{solver_traj}{}

\item{successRatio}{}
}
\value{

}
\description{
Title
}
