% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuadraticModelCoeff.R
\name{getQuadraticModelCoeff}
\alias{getQuadraticModelCoeff}
\title{Title}
\usage{
getQuadraticModelCoeff(solver_traj, plot = FALSE)
}
\arguments{
\item{solver_traj}{}

\item{plot}{}
}
\value{

}
\description{
Title
}
