% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getViolationAreaStats.R
\name{getVioPlat}
\alias{getVioPlat}
\title{Title}
\usage{
getVioPlat(solver_traj, VG_stats, Plat_start_stats, Plat_stats)
}
\arguments{
\item{solver_traj}{}

\item{VG_stats}{}

\item{Plat_start_stats}{}

\item{Plat_stats}{}
}
\value{

}
\description{
Title
}
