% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInstanceLoader.R
\name{load_netgen.cluster_instances}
\alias{load_netgen.cluster_instances}
\title{Title}
\usage{
load_netgen.cluster_instances(assembled, n.dim, n.cluster)
}
\arguments{
\item{assembled}{}

\item{n.dim}{}

\item{n.cluster}{}
}
\value{

}
\description{
Title
}
