% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInstanceLoader.R
\name{load_netgen.grid_instances}
\alias{load_netgen.grid_instances}
\title{Title}
\usage{
load_netgen.grid_instances(assembled, n.dim)
}
\arguments{
\item{assembled}{}

\item{n.dim}{}
}
\value{

}
\description{
Title
}
