% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInstanceLoader.R
\name{load_tsplib_instances}
\alias{load_tsplib_instances}
\title{Title}
\usage{
load_tsplib_instances(path_to_testdata, path_to_salesperson)
}
\arguments{
\item{path_to_testdata}{}

\item{path_to_salesperson}{}
}
\value{

}
\description{
Title
}
