% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInstanceLoader.R
\name{make_generator_instances}
\alias{make_generator_instances}
\title{Title}
\usage{
make_generator_instances(
  assembled,
  n.dim,
  n.cluster,
  path_to_testdata,
  path_to_salesperson,
  ...
)
}
\arguments{
\item{assembled}{}

\item{n.dim}{}

\item{n.cluster}{}

\item{path_to_testdata}{}

\item{path_to_salesperson}{}

\item{...}{}
}
\value{

}
\description{
Title
}
