% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConvergenceSpeed.R
\name{makeConvSpeed_2_strict_plot}
\alias{makeConvSpeed_2_strict_plot}
\title{Title}
\usage{
makeConvSpeed_2_strict_plot(solver_traj, solver_traj_wo_plat)
}
\arguments{
\item{solver_traj}{}

\item{solver_traj_wo_plat}{}
}
\value{

}
\description{
Title
}
