% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAreaIncLB.R
\name{makeIncLB_area_plot}
\alias{makeIncLB_area_plot}
\title{Title}
\usage{
makeIncLB_area_plot(solver_traj, lb, IncLB_stat)
}
\arguments{
\item{solver_traj}{}

\item{lb}{}

\item{IncLB_stat}{}
}
\value{

}
\description{
Title
}
