% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getViolationAreaStats.R
\name{makePlateauVio_plot}
\alias{makePlateauVio_plot}
\title{Title}
\usage{
makePlateauVio_plot(vio_stats, solver_traj, VGstats)
}
\arguments{
\item{vio_stats}{}

\item{solver_traj}{}

\item{VGstats}{}
}
\value{

}
\description{
Title
}
