% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_monitor_feature.R
\name{makeStats}
\alias{makeStats}
\title{Title}
\usage{
makeStats(name, stat_ls, stat_flag = T)
}
\arguments{
\item{name}{}

\item{stat_ls}{}

\item{stat_flag}{}
}
\value{

}
\description{
Title
}
