% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectHelper.R
\name{min_max_memory}
\alias{min_max_memory}
\title{Creates Memory of original values of solver object}
\usage{
min_max_memory(solvertraj, classAdd)
}
\arguments{
\item{solvertraj}{}

\item{classAdd}{}
}
\value{

}
\description{
Creates Memory of original values of solver object
}
