% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scalerOwn.scale_other}
\alias{scalerOwn.scale_other}
\title{Method for res_eax Object}
\usage{
\method{scalerOwn}{scale_other}(solvertraj, ...)
}
\arguments{
\item{solvertraj}{}

\item{...}{}
}
\value{

}
\description{
Method for res_eax Object
}
