% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scalerOwn.scale_Scalar}
\alias{scalerOwn.scale_Scalar}
\title{Method for Scalar values}
\usage{
\method{scalerOwn}{scale_Scalar}(LB, min_inc, max_inc, ...)
}
\arguments{
\item{LB}{}

\item{min_inc}{}

\item{max_inc}{}
}
\value{

}
\description{
Method for Scalar values
}
