% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSolverTrajLen.R
\name{solvertraj_len}
\alias{solvertraj_len}
\title{SolverTraj Length and Stats}
\usage{
solvertraj_len(solvertraj, which_one)
}
\arguments{
\item{solvertraj}{}

\item{which_one}{}
}
\value{

}
\description{
SolverTraj Length and Stats
}
