% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSplit.R
\name{train_splitter}
\alias{train_splitter}
\title{Splitter Function}
\usage{
train_splitter(split_data, valid_wanted = FALSE, train_split, seed, ...)
}
\arguments{
\item{split_data}{}

\item{valid_wanted}{}

\item{train_split}{}

\item{seed}{}

\item{...}{}
}
\value{

}
\description{
Splitter Function
}
