% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_monitor_feature.R
\name{try.getFeatureSet}
\alias{try.getFeatureSet}
\title{Title}
\usage{
try.getFeatureSet(black.list, instance)
}
\arguments{
\item{black.list}{}

\item{instance}{}
}
\value{

}
\description{
Title
}
